+++ NEWS.md
@@ -1,3 +1,21 @@
+# CHANGES IN xfun VERSION 0.22
+
+## NEW FEATURES
+
+- `relative_path()` is vectorized now.
+
+- Added a new function `retry()` to retry calling a function for a number of times in case of errors.
+
+- Added a new function `sort_file()`, which is a shorthand for `process_file(fun = sort)` to sort the lines in a text file.
+
+## MAJOR CHANGES
+
+- The argument `FUN` was renamed to `fun` in `process_file()`.
+
+## MINOR CHANGES
+
+- The `timeout` option in `options()` is set to 3600 seconds when it takes the default value of 60 seconds, which may not be enough for downloading large files (thanks, @matthewgson, yihui/tinytex#286).
+
 # CHANGES IN xfun VERSION 0.21
 
 ## NEW FEATURES